fix(web): rename SkipToContent to SkipToContentLink
Avoids using the same name as the underlying PF component to reduce
ambiguity for developers.
